chore: group skip-dir related elements
