+++ pkg/result/processors/skip_dirs_test.go
@@ -2,6 +2,8 @@ package processors
 
 import (
 	"path/filepath"
+	"regexp"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -55,6 +57,38 @@ func Test_absDirs(t *testing.T) {
 
 			assert.Equal(t, test.expected, results)
 		})
+	}
+}
+
+func Test_createPathRegex(t *testing.T) {
+	matches := [][]string{
+		{"dir"},
+		{"root", "dir"},
+		{"root", "dir", "subdir"},
+		{"dir", "subdir"},
+	}
+
+	noMatches := [][]string{
+		{"nodir"},
+		{"dirno"},
+		{"root", "dirno"},
+		{"root", "nodir"},
+		{"root", "dirno", "subdir"},
+		{"root", "nodir", "subdir"},
+		{"dirno", "subdir"},
+		{"nodir", "subdir"},
+	}
+
+	for _, sep := range []rune{'/', '\\'} {
+		exp := regexp.MustCompile(createPathRegex("dir", sep))
+
+		for _, m := range matches {
+			assert.Regexp(t, exp, strings.Join(m, string(sep)))
+		}
+
+		for _, m := range noMatches {
+			assert.NotRegexp(t, exp, strings.Join(m, string(sep)))
+		}
 	}
 }
